+++ app/src/main/java/com/work/guaishouxingqiu/aboutball/venue/activity/VenueDetailsActivity.java
@@ -222,8 +222,8 @@ public class VenueDetailsActivity extends BaseActivity<VenueDetailsPresenter> im
                 mDetailsBean.areaForDetailList.size() > mSelectorTabPosition) {
             Bundle bundle = new Bundle();
             bundle.putInt(ARouterConfig.Key.POSITION, mCalendarPosition);
-            bundle.putInt(ARouterConfig.Key.AREA_ID, mDetailsBean.areaForDetailList.get(mSelectorTabPosition).areaId);
-            bundle.putInt(ARouterConfig.Key.STADIUM_ID, mStadiumId);
+            bundle.putLong(ARouterConfig.Key.AREA_ID, mDetailsBean.areaForDetailList.get(mSelectorTabPosition).areaId);
+            bundle.putLong(ARouterConfig.Key.STADIUM_ID, mStadiumId);
             IntentData.get().putData(mDetailsBean.areaForDetailList.get(mSelectorTabPosition).calendarListForAreaList);
             ARouterIntent.startActivity(ARouterConfig.Path.ACTIVITY_VENUE_BOOKING,bundle);
         }
